CF like method to return a mutable path that is in the shape of a rounded rectangle
corresponding methods to draw rounded rectangles with a solid fill color and with a gradient.

abstracted gradient drawing into a method that will draw to a context you specify. also used a more primitive way of dealing with the colors

UIColor additions to deal with getting of converted RGB components in the absence of colorspace conversion mechanics